refactored namespaces in FaceRecognition
