+++ src/Isidore.MagicMirror.ImageProcessing/FaceRecognition/Classifiers/IFaceClassifier.cs
@@ -1,8 +1,9 @@
 using System.Collections.Generic;
+using Isidore.MagicMirror.ImageProcessing.FaceRecognition.Models;
 
 namespace Isidore.MagicMirror.ImageProcessing.FaceRecognition.Classifiers
 {
-   public interface IFaceClassifier<TImage>
+    public interface IFaceClassifier<TImage>
     {
         IEnumerable<Area> DetectAllFaces(TImage image);
