+++ src/Isidore.MagicMirror.ImageProcessing/FaceRecognition/Models/RecognitionResult.cs
@@ -1,4 +1,4 @@
-namespace Isidore.MagicMirror.ImageProcessing.FaceRecognition
+namespace Isidore.MagicMirror.ImageProcessing.FaceRecognition.Models
 {
     public class RecognitionResult<T>
     {
