+++ src/Isidore.MagicMirror.ImageProcessing/FaceRecognition/Services/FisherFaceService.cs
@@ -5,10 +5,12 @@ using System.Linq;
 using System.Threading.Tasks;
 using Isidore.MagicMirror.ImageProcessing.FaceRecognition.Classifiers;
 using Isidore.MagicMirror.ImageProcessing.Helpers;
+using Isidore.MagicMirror.ImageProcessing.FaceRecognition.Models;
+using Isidore.MagicMirror.ImageProcessing.FaceRecognition.Services;
 using OpenCvSharp;
 using OpenCvSharp.Face;
 
-namespace Isidore.MagicMirror.ImageProcessing.FaceRecognition
+namespace Isidore.MagicMirror.ImageProcessing.FaceRecognition.Services
 {
     public class FisherFaceService : IFaceRecognitionService<Mat>
     {
